pyproject.toml
[tool.poetry]
name = "list-products"
version = "0.1.0"
description = ""
authors = ["lu-css <dossantossilvalucas963@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
chromadb = "^0.4.24"
python-dotenv = "^1.0.1"
zope-interface = "^6.2"
flask = "^3.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
